State Comparison
Using Tkinter, create a modern GUI which compares multiple attributes such as population, sex ratio, literacy rate, population density, birth rate, death rate and the fertility rate of all the states and union territories in India.
The GUI application is an easy-to-understand tool which has taken data from the RBI website and compared the various attributes for different states and union territories. It also shows the average rate for the country for each of the attributes via the plots. When the file is run, it looks like an app built for the laptop.